# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditCategory": {
        "enum": [
          "AUTH",
          "EXAM",
          "SESSION",
          "CONNECTION",
          "SYSTEM"
        ],
        "title": "AuditCategory",
        "type": "string"
      },
      "AuditEventOut": {
        "description": "One append-only trail entry.\n\nTwo timestamps, deliberately. ``occurred_at`` is what the reporting client\nclaimed; ``recorded_at`` is when the server accepted it. Keeping both means\nclock skew - or a deliberately altered client clock - is visible in the\ndata instead of silently corrupting the ordering of the timeline.",
        "properties": {
          "actorLabel": {
            "title": "Actorlabel",
            "type": "string"
          },
          "actorType": {
            "$ref": "#/components/schemas/SubjectType"
          },
          "category": {
            "$ref": "#/components/schemas/AuditCategory"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "event": {
            "$ref": "#/components/schemas/AuditEventType"
          },
          "examId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Examid"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "machineId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Machineid"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "recordedAt": {
            "format": "date-time",
            "title": "Recordedat",
            "type": "string"
          },
          "sessionId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessionid"
          },
          "severity": {
            "$ref": "#/components/schemas/AuditSeverity"
          },
          "studentId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Studentid"
          }
        },
        "required": [
          "id",
          "event",
          "category",
          "severity",
          "detail",
          "occurredAt",
          "recordedAt",
          "actorType",
          "actorLabel"
        ],
        "title": "AuditEventOut",
        "type": "object"
      },
...
      "AuditSeverity": {
        "enum": [
          "INFO",
          "WARNING",
          "CRITICAL"
        ],
        "title": "AuditSeverity",
        "type": "string"
      },
...
      "ComputerOut": {
        "description": "A workstation.\n\nDeliberately has no ``student_id``. Seating is a fact about a candidate\nsitting a particular exam, recorded on the session - a lab hosts a\ndifferent cohort every hour, so a permanent binding cannot hold.",
        "properties": {
          "connection": {
            "$ref": "#/components/schemas/ConnectionState"
          },
          "enrolledAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enrolledat"
          },
          "hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hostname"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "labId": {
            "format": "uuid",
            "title": "Labid",
            "type": "string"
          },
          "lastHeartbeatAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastheartbeatat"
          },
          "machineId": {
            "title": "Machineid",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "machineId",
          "labId",
          "position",
          "connection"
        ],
        "title": "ComputerOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "MachineCredential": {
        "properties": {
          "labId": {
            "format": "uuid",
            "title": "Labid",
            "type": "string"
          },
          "machineId": {
            "title": "Machineid",
            "type": "string"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "machineId",
          "secret",
          "labId"
        ],
        "title": "MachineCredential",
        "type": "object"
      },
      "MachineEnrolRequest": {
        "description": "One-time enrolment, performed once per workstation by an administrator.\n\nA workstation is not a user: it cannot own an exam, sit in a roster, or\nreset a password. Modelling it as a user row would mean every query\ntouching users has to remember to exclude machines.",
        "properties": {
          "enrolmentToken": {
            "title": "Enrolmenttoken",
            "type": "string"
          },
          "hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hostname"
          },
          "machineId": {
            "title": "Machineid",
            "type": "string"
          }
        },
        "required": [
          "enrolmentToken",
          "machineId"
        ],
        "title": "MachineEnrolRequest",
        "type": "object"
      },
      "MachineToken": {
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "labId": {
            "format": "uuid",
            "title": "Labid",
            "type": "string"
          },
          "machineId": {
            "title": "Machineid",
            "type": "string"
          },
          "serverTime": {
            "format": "date-time",
            "title": "Servertime",
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "expiresAt",
          "serverTime",
          "machineId",
          "labId"
        ],
        "title": "MachineToken",
        "type": "object"
      },
      "MachineTokenRequest": {
        "properties": {
          "machineId": {
            "title": "Machineid",
            "type": "string"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "machineId",
          "secret"
        ],
        "title": "MachineTokenRequest",
        "type": "object"
      },
...
      "Page_AuditEventOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEventOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[AuditEventOut]",
        "type": "object"
      },
...
      "Page_StudentOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudentOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[StudentOut]",
        "type": "object"
      },
...
      "RefreshRequest": {
        "properties": {
          "refreshToken": {
            "title": "Refreshtoken",
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "StudentCreate": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "fullName": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Fullname",
            "type": "string"
          },
          "program": {
            "title": "Program",
            "type": "string"
          },
          "registrationNo": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Registrationno",
            "type": "string"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "semester": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Semester",
            "type": "integer"
          }
        },
        "required": [
          "registrationNo",
          "fullName",
          "email",
          "program",
          "semester",
          "section"
        ],
        "title": "StudentCreate",
        "type": "object"
      },
...
      "StudentOut": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "fullName": {
            "title": "Fullname",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "program": {
            "title": "Program",
            "type": "string"
          },
          "registrationNo": {
            "title": "Registrationno",
            "type": "string"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "semester": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Semester",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/StudentStatus"
          }
        },
        "required": [
          "id",
          "registrationNo",
          "fullName",
          "email",
          "program",
          "semester",
          "section",
          "status"
        ],
        "title": "StudentOut",
        "type": "object"
      },
...
      "StudentStatus": {
        "enum": [
          "ACTIVE",
          "BLOCKED"
        ],
        "title": "StudentStatus",
        "type": "string"
      },
      "StudentUpdate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "fullName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fullname"
          },
          "program": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "semester": {
            "anyOf": [
              {
                "maximum": 12.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StudentStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "StudentUpdate",
        "type": "object"
      },
...
      "TokenPair": {
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "refreshToken": {
            "title": "Refreshtoken",
            "type": "string"
          },
          "serverTime": {
            "format": "date-time",
            "title": "Servertime",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresAt",
          "serverTime",
          "user"
        ],
        "title": "TokenPair",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Phase 2 contract. Handlers return static examples; the shapes, status codes and operation ids are final and generate the frontend client.",
    "summary": "Server-authoritative examination platform for physical computer labs.",
    "title": "Exam Control API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/audit": {
      "get": {
        "description": "Append-only trail. There is deliberately no delete endpoint.",
        "operationId": "listAuditEvents",
        "parameters": [
          {
            "in": "query",
            "name": "examId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Examid"
            }
          },
          {
            "in": "query",
            "name": "studentId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Studentid"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AuditSeverity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AuditCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AuditEventOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit Events",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Human sign-in. The response carries ``server_time`` so the client can\nestablish its clock offset before anything is timed.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/machine/enrol": {
      "post": {
        "description": "One-time workstation enrolment, performed by an administrator.\n\nReturns the machine secret exactly once; it is stored hashed.",
        "operationId": "enrolMachine",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MachineEnrolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineCredential"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enrol Machine",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/machine/token": {
      "post": {
        "description": "Exchanges a machine credential for a short-lived token carrying\n``subject_type=machine``. Such a token may only reach heartbeat and event\nendpoints.",
        "operationId": "machineToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MachineTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineToken"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Machine Token",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/labs/{lab_id}/computers": {
      "get": {
        "description": "Workstations and their derived liveness.\n\nNo candidate is attached here: seating belongs to the session, because a\nlab hosts a different cohort every hour.",
        "operationId": "listLabComputers",
        "parameters": [
          {
            "in": "path",
            "name": "lab_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lab Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ComputerOut"
                  },
                  "title": "Response Listlabcomputers",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Lab Computers",
        "tags": [
          "directory"
        ]
      }
    },
...
    "/api/v1/students": {
      "get": {
        "operationId": "listStudents",
        "parameters": [
          {
            "description": "Matches name or registration number.",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Matches name or registration number.",
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "section",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section"
            }
          },
          {
            "in": "query",
            "name": "semester",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 12,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semester"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StudentOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Students",
        "tags": [
          "directory"
        ]
      },
      "post": {
        "operationId": "createStudent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Student",
        "tags": [
          "directory"
        ]
      }
    },
    "/api/v1/students/{student_id}": {
      "patch": {
        "operationId": "updateStudent",
        "parameters": [
          {
            "in": "path",
            "name": "student_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Student Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Student",
        "tags": [
          "directory"
        ]
      }
    },
...
  }
}